Universal Version 1.21
The system should now work correctly for every log in case in terms of
voting and being able to post comments.

Content and formatting has been improved on the add new question and
return question pages by centre aligning them.

A feature has been added where the system asks the user if they want to
log in straight after registering and if not leaves them on the same
page, but their details are still registered. Validation has been
applied to fix the registry error.