UV 2.8
Error message when resetting password wrongly now disappears on the
second try. The most recent comments to a question appear first. The
query now deletes any responses related to a question before deleting
that question, and the database copy has cascade on it (working master
version). The links from the user's page to their question no longer
causes an error. The downvote and upvote system should now work, and
return the user to the original question page.